Move aladin related files to dedicated folder scripts/ci/aladdin
